+++ webapp/app/components/pagination-nav/component.js
@@ -3,6 +3,9 @@ import Ember from "ember";
 export default Ember.Component.extend({
   page: 1,
   has_next: false,
+  num_pages: null,
+
+  has_last: Ember.computed.notEmpty('num_pages'),
 
   has_prev: Ember.computed.gt("page", 1),
 
@@ -17,6 +20,10 @@ export default Ember.Component.extend({
 
     prev_page() {
       this.decrementProperty("page");
-    }
+    },
+
+    last_page() {
+      this.set('page', this.get('num_pages'));
+    },
   }
 });