+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Version 2.7.0
 
+* You can now jump to the "last" page in the session tests view
 * Added monitoring compatible with Prometheus (both natively via `/metrics` and the postgres exporter on port 9187)
 * Tracebacks and errors are significantly improved, and now support collapsing of `self` attributes, copying the full exception to the clipboard and more
 * Backslash now supports a newer way of uploading tracebacks via streaming uploads, improving performance on large tracebacks